Revamp of the build system
Due to a lot of complications with sbt and plugins in subprojects,
libinteractive is now a totally separate project. This means that the
way the backend is built needed some updating. This change centralizes
and improves everything, and does not copy the proguard-compiled files
into the bin/ directory in omegaUp since that will be omegaUp's
responsibility.