Refactor services
The goal of this refactoring is to cleanly define interfaces and
services and to decouple the services from one another. This also
renames grader.Grader => grader.Validator (to better match the database
schema and its purpose), grader.Manager => grader.Grader (to be
consistent with the Runner and Broadcaster), and move the HTTP server
from Grader to Service (since it also depends on Broadcaster).