Allow per-request logging through the Context
Introduce an OverrideLogger that copies log entries to a memory buffer
and can be later accessed. This is in preparation for #9 and allowed a
massive cleanup in the tests. Now log statements do not clutter the
screen while tests are running, and are only displayed after the test is
run, only if the test case fails, and together with the failure
information.