Build improvements
This change enables the use of sbt-assembly in addition to proguard.
sbt-assembly takes way less time and memory and can be used to produce
binaries in the super weak compilation server. This also uses git
version numbers for the .jar versions, so each build will have a unique
version number.