Fixed a race
If network conditions are bad (which we expect them to be), a run might
be submitted, but the acknowledgement with the guid might arrive later
than the WebSocket notification that the run has finished running. In
that case, store the run details in an orphan map where it can be
processed after it is inserted.